+++ 1.18/Fabric/src/main/java/fuzs/respawninganimals/mixin/NaturalSpawnerFabricMixin.java
@@ -4,13 +4,11 @@ import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import fuzs.respawninganimals.world.entity.SpawnTypeMob;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.MobSpawnType;
 import net.minecraft.world.level.NaturalSpawner;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
 @Mixin(NaturalSpawner.class)
 abstract class NaturalSpawnerFabricMixin {
@@ -30,16 +28,5 @@ abstract class NaturalSpawnerFabricMixin {
         } else {
             return false;
         }
-    }
-
-    @ModifyVariable(
-            method = "spawnMobsForChunkGeneration", at = @At(
-            value = "INVOKE",
-            target = "Lnet/minecraft/world/entity/Mob;finalizeSpawn(Lnet/minecraft/world/level/ServerLevelAccessor;Lnet/minecraft/world/DifficultyInstance;Lnet/minecraft/world/entity/MobSpawnType;Lnet/minecraft/world/entity/SpawnGroupData;Lnet/minecraft/nbt/CompoundTag;)Lnet/minecraft/world/entity/SpawnGroupData;"
-    )
-    )
-    private static Entity spawnMobsForChunkGeneration(Entity entity) {
-        ((SpawnTypeMob) entity).respawninganimals$setSpawnType(MobSpawnType.CHUNK_GENERATION);
-        return entity;
     }
 }
