+++ 1.18/Fabric/src/main/resources/fabric.mod.json
@@ -30,7 +30,8 @@
   },
 
   "mixins": [
-    "${modId}.common.mixins.json"
+    "${modId}.common.mixins.json",
+    "${modId}.fabric.mixins.json"
   ],
 
   "depends": {
