+++ 1.18/Forge/src/main/java/fuzs/respawninganimals/mixin/NaturalSpawnerForgeMixin.java
@@ -1,0 +1,25 @@
+package fuzs.respawninganimals.mixin;
+
+import fuzs.respawninganimals.world.entity.SpawnTypeMob;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.level.NaturalSpawner;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
+
+@Mixin(NaturalSpawner.class)
+abstract class NaturalSpawnerForgeMixin {
+
+    @ModifyVariable(
+            method = "spawnCategoryForPosition(Lnet/minecraft/world/entity/MobCategory;Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/level/chunk/ChunkAccess;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/NaturalSpawner$SpawnPredicate;Lnet/minecraft/world/level/NaturalSpawner$AfterSpawnCallback;)V",
+            at = @At(
+                    value = "INVOKE",
+                    target = "Lnet/minecraft/world/entity/Mob;finalizeSpawn(Lnet/minecraft/world/level/ServerLevelAccessor;Lnet/minecraft/world/DifficultyInstance;Lnet/minecraft/world/entity/MobSpawnType;Lnet/minecraft/world/entity/SpawnGroupData;Lnet/minecraft/nbt/CompoundTag;)Lnet/minecraft/world/entity/SpawnGroupData;"
+            )
+    )
+    private static Mob spawnCategoryForPosition(Mob mob) {
+        ((SpawnTypeMob) mob).respawninganimals$setSpawnType(MobSpawnType.NATURAL);
+        return mob;
+    }
+}
